+++ src/telegram/open-ai.service.ts
@@ -19,7 +19,7 @@ export class OpenAiService {
     messages: ChatCompletionMessageParam[],
   ): Promise<ChatCompletionMessageParam[]> {
     const completion = await this.openai.chat.completions.create({
-      model: 'gpt-3.5-turbo',
+      model: process.env.OPENAI_GPT_MODEL,
       messages,
     })